requirements.txt
aiohttp>=3.9,<4
bcrypt>=4.1,<5
fastapi>=0.115,<1
gradio>=5,<6
gradio-imageslider>=0.0.20,<0.1
numpy>=1.26,<3
Pillow>=10,<12
python-dotenv>=1.0,<2
requests>=2.31,<3
